requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dicttoxml==1.7.16
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
greenlet==3.3.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.3
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.20.0